improve handling of categorical data, add nearby word tokens
use lightgbm's handling of categorical data instead of one-hot encoding. This
shrinking of the feature space means I could test out encoding the previous,
following, and parse-tree-root tokens as well as the current tokens. I have to
run on the full data set but it looks like this could be a big performance jump!
Possibly around .007 AUC. Ok so much for going to bed early...